gyp: export GYP_ENGINE/GYP_ENGINE_VERSION
`gyp.js` has a bit different semantics when it comes to placement of the
output files. To be able to build node `node.gyp` will have to be
modified to account for this. In such case, additional dependence on
particular GYP engine will be needed.

`GYP_ENGINE` and `GYP_ENGINE_VERSION` are two variables that will help
maintainers to add such dependence and make the project universally
buildable.